backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Stripe
stripe==11.4.1

# Geo
geopy==2.4.1

# PDF generation
weasyprint==63.1
jinja2==3.1.5

# File storage (S3/R2)
boto3==1.36.4

# Background tasks
apscheduler==3.10.4

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.4.0

# Error tracking
sentry-sdk[fastapi]==2.52.0

# HTTP client
httpx==0.28.1

# Email (Resend)
resend==2.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.3.1
aiosqlite==0.20.0
